4x4 Tic Tac Toe with Deep learning
This game has single-player and multiplayer modes. The single-player mode encompasses deep learning algorithms which gets more vigorous as more games have culminated and more data is integrated into the solutions file.
There are 16 built-in solutions that are provided at default. The arduousness of the game increases with the incrementation in the number of solutions.
The high score exhibits all the best players of the game.
There is account engenderment as well, it's not required, but it integrates onto the features.

This program was developed in the C++ language.